Moved main screen logic to SearchScreen.kt, added preliminary code for bottom navbar, added drawables for the bar
