+++ app/build.gradle
@@ -59,7 +59,7 @@ dependencies {
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
     implementation 'androidx.navigation:navigation-runtime-ktx:2.5.3'
-    implementation "androidx.navigation:navigation-compose:2.4.0-alpha10"
+    implementation "androidx.navigation:navigation-compose:2.4.0-beta02"
     implementation 'androidx.compose.material:material:1.0.5'
     implementation "androidx.navigation:navigation-compose:2.4.0-alpha10"
     implementation "androidx.compose.material:material-icons-core:1.0.5"
