+++ app/src/main/java/com/example/sampleappcompose/MainActivity.kt
@@ -12,8 +12,11 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.material.BottomNavigation
+import androidx.compose.material.BottomNavigationItem
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
@@ -23,6 +26,10 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -41,115 +48,41 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-@Composable
-fun SearchApp() {
-    val logs = remember { mutableStateListOf<String>() }
-    val apiResponse = remember { mutableStateOf("") }
-    val requestData = remember { mutableStateOf("") }
-    val coroutineScope = rememberCoroutineScope()
 
-    Column(
-        modifier = Modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
+@Composable
+fun BottomNavigationBar() {
+    val items = listOf(
+        Screens.Home,
+        Screens.Search,
+        Screens.Info
+    )
+    BottomNavigation(
+        backgroundColor = Color.Black,
+        contentColor = Color.White
     ) {
-        Form(
-            requestData = requestData.value,
-            onValueChange = { newValue -> requestData.value = newValue },
-            onSubmit = {
-                coroutineScope.launch {
-                    launchApiCall(requestData.value, apiResponse, logs)
+        items.forEach { item ->
+            BottomNavigationItem(
+                icon = { Icon(painterResource(id = item.icon), contentDescription = item.title) },
+                label = { Text(text = item.title) },
+                selectedContentColor = Color.White,
+                unselectedContentColor = Color.White.copy(0.4f),
+                alwaysShowLabel = true,
+                selected = false,
+                onClick = {
+                    /* TODO */
                 }
-            }
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        ResponseView(response = apiResponse.value)
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        LogsView(logs = logs)
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun Form(requestData: String, onValueChange: (String) -> Unit, onSubmit: () -> Unit) {
-    Column(
-        modifier = Modifier.padding(16.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        androidx.compose.material3.TextField(
-            value = requestData,
-            onValueChange = onValueChange,
-            label = { Text("Enter movie title") },
-            modifier = Modifier.fillMaxWidth()
-        )
-
-        androidx.compose.material3.Button(
-            onClick = onSubmit,
-            modifier = Modifier.align(Alignment.End)
-        ) {
-            Text(text = "Submit")
+            )
         }
     }
 }
-
+@Preview(showBackground = true)
 @Composable
-fun ResponseView(response: String) {
-    Column(
-        modifier = Modifier.padding(16.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        Text(text = "API Response:")
-        Text(text = response)
-    }
+fun BottomNavigationBarPreview() {
+    BottomNavigationBar()
 }
 
-@Composable
-fun LogsView(logs: List<String>) {
-    Column(
-        modifier = Modifier.padding(16.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        Text(text = "API Logs:")
-        LazyColumn {
-            items(items = logs) { log -> Text(text = log) }
-        }
-    }
-}
 
-data class ApiResponse(val Title: String, val Year: String, val imdbRating: String)
 
-interface ApiService {
-    @GET("/")
-    suspend fun makeApiCall(@Query("apikey") apikey: String, @Query("t") title: String): ApiResponse
-}
-
-suspend fun makeApiCall(movieTitle: String): ApiResponse {
-    val retrofit = Retrofit.Builder()
-        .baseUrl("http://www.omdbapi.com/") // OMDb API
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-
-    val apiService = retrofit.create(ApiService::class.java)
-
-    return withContext(Dispatchers.IO) {
-        apiService.makeApiCall("6b938676", movieTitle)
-    }
-}
-
-suspend fun launchApiCall(movieTitle: String, apiResponse: MutableState<String>, logs: MutableList<String>) {
-    try {
-        print("Test")
-        val response = makeApiCall(movieTitle)
-        apiResponse.value = "Title: ${response.Title} \n Year: ${response.Year} \n IMDb Rating: ${response.imdbRating}"
-        logs.add("API call made successfully.")
-    } catch (e: Exception) {
-        logs.add("API call failed: ${e.localizedMessage}")
-    }
-}
